there is no need for .Win32.cs and .Windows.cs file anymore
